Rename safety concept to safety considerations
Fixes #5508
