+++ api/fixtures/categories.yml
@@ -7,7 +7,7 @@ App\Entity\Category:
     numberingStyle: 1
     rootContentNode: '@columnLayout2'
     preferredContentTypes:
-      - '@contentTypeSafetyConcept'
+      - '@contentTypeSafetyConsiderations'
   category2:
     camp: '@camp1'
     short: LP
