+++ api/tests/Api/ContentNodes/ContentNode/ListContentNodesTest.php
@@ -45,7 +45,7 @@ class ListContentNodesTest extends ECampApiTestCase {
             ['href' => $this->getIriFor('columnLayout2campPrototype')],
             ['href' => $this->getIriFor('singleText1')],
             ['href' => $this->getIriFor('singleText2')],
-            ['href' => $this->getIriFor('safetyConcept1')],
+            ['href' => $this->getIriFor('safetyConsiderations1')],
             ['href' => $this->getIriFor('materialNode1')],
             ['href' => $this->getIriFor('materialNode2')],
             ['href' => $this->getIriFor('storyboard1')],
@@ -75,7 +75,7 @@ class ListContentNodesTest extends ECampApiTestCase {
             ['href' => $this->getIriFor('columnLayout3')],
             ['href' => $this->getIriFor('singleText1')],
             ['href' => $this->getIriFor('singleText2')],
-            ['href' => $this->getIriFor('safetyConcept1')],
+            ['href' => $this->getIriFor('safetyConsiderations1')],
             ['href' => $this->getIriFor('materialNode1')],
             ['href' => $this->getIriFor('storyboard1')],
             ['href' => $this->getIriFor('storyboard2')],
