+++ api/tests/Api/ContentTypes/ReadContentTypeTest.php
@@ -11,7 +11,7 @@ use App\Tests\Api\ECampApiTestCase;
 class ReadContentTypeTest extends ECampApiTestCase {
     public function testGetSingleContentTypeIsAllowedForAnonymousUser() {
         /** @var ContentType $contentType */
-        $contentType = static::getFixture('contentTypeSafetyConcept');
+        $contentType = static::getFixture('contentTypeSafetyConsiderations');
         static::createBasicClient()->request('GET', '/content_types/'.$contentType->getId());
         $this->assertResponseStatusCodeSame(200);
         $this->assertJsonContains([
@@ -28,7 +28,7 @@ class ReadContentTypeTest extends ECampApiTestCase {
 
     public function testGetSingleContentTypeIsAllowedForLoggedInUser() {
         /** @var ContentType $contentType */
-        $contentType = static::getFixture('contentTypeSafetyConcept');
+        $contentType = static::getFixture('contentTypeSafetyConsiderations');
         static::createClientWithCredentials()->request('GET', '/content_types/'.$contentType->getId());
         $this->assertResponseStatusCodeSame(200);
         $this->assertJsonContains([
