+++ pdf/src/campPrint/scheduleEntry/contentNode/ContentNode.vue
@@ -15,7 +15,7 @@ import LearningObjectives from './LearningObjectives.vue'
 import LearningTopics from './LearningTopics.vue'
 import Storyboard from './Storyboard.vue'
 import Notes from './Notes.vue'
-import SafetyConcept from './SafetyConcept.vue'
+import SafetyConsiderations from './SafetyConsiderations.vue'
 import Material from './Material.vue'
 import Storycontext from './Storycontext.vue'
 
@@ -38,7 +38,7 @@ export default {
         LearningTopics,
         Storyboard,
         Notes,
-        SafetyConcept,
+        SafetyConsiderations,
         Material,
         Storycontext,
       }[this.contentTypeName]
