+++ pdf/src/renderer/__tests__/fullCampStoreContent.json
@@ -1190,7 +1190,7 @@
     "position": 1,
     "instanceName": null,
     "id": "dbed0f43dd70",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/193c3491fa23"
     },
@@ -1516,7 +1516,7 @@
     "position": 1,
     "instanceName": null,
     "id": "95cd1a28febb",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/13d74b9d52b8"
     },
@@ -2038,7 +2038,7 @@
     "position": 1,
     "instanceName": null,
     "id": "d5114af3e81e",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/e460f03e79b4"
     },
@@ -2320,7 +2320,7 @@
     "position": 1,
     "instanceName": null,
     "id": "7fe23d0fc00f",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/d81a9dae4316"
     },
@@ -2722,7 +2722,7 @@
     "position": 1,
     "instanceName": null,
     "id": "89a401571d37",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/3ddaa17f2649"
     },
@@ -3073,7 +3073,7 @@
     "position": 1,
     "instanceName": null,
     "id": "49d09f48b287",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/475f7b34581a"
     },
@@ -3484,7 +3484,7 @@
     "position": 1,
     "instanceName": null,
     "id": "21bca97eadd7",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/2b3f79233387"
     },
@@ -3927,7 +3927,7 @@
     "position": 1,
     "instanceName": null,
     "id": "e5c2b710898d",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/7792c9c855ce"
     },
@@ -4275,7 +4275,7 @@
     "position": 1,
     "instanceName": null,
     "id": "1f648080582d",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/b46ab3cde287"
     },
@@ -4718,7 +4718,7 @@
     "position": 1,
     "instanceName": null,
     "id": "22d8c12d7df3",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/d4fe28437432"
     },
@@ -5255,7 +5255,7 @@
     "position": 1,
     "instanceName": null,
     "id": "84b40e43160a",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/280031715953"
     },
@@ -5710,7 +5710,7 @@
     "position": 1,
     "instanceName": null,
     "id": "b3f5cbd532da",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/f9ba3a735450"
     },
@@ -6191,7 +6191,7 @@
     "position": 1,
     "instanceName": null,
     "id": "0fada08fb7c1",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/8e15971caab8"
     },
@@ -6527,7 +6527,7 @@
     "position": 1,
     "instanceName": null,
     "id": "f2f21f7b5963",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/3a9614e39661"
     },
@@ -6875,7 +6875,7 @@
     "position": 1,
     "instanceName": null,
     "id": "d0448480da8b",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/aff5fa86266e"
     },
@@ -7318,7 +7318,7 @@
     "position": 1,
     "instanceName": null,
     "id": "a7a4a7ebefe6",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/5015df0b6259"
     },
@@ -7711,7 +7711,7 @@
     "position": 1,
     "instanceName": null,
     "id": "ae634c4157a0",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/69fa08a77f38"
     },
@@ -8109,7 +8109,7 @@
     "position": 1,
     "instanceName": null,
     "id": "73ce151126d3",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/c2eae59077c5"
     },
@@ -8549,7 +8549,7 @@
     "position": 1,
     "instanceName": null,
     "id": "381c06d8ac17",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/78c9196140b7"
     },
@@ -8969,7 +8969,7 @@
     "position": 1,
     "instanceName": null,
     "id": "9a405af3cf69",
-    "contentTypeName": "SafetyConcept",
+    "contentTypeName": "SafetyConsiderations",
     "root": {
       "href": "/content_node/column_layouts/985356416a4b"
     },
@@ -9925,7 +9925,7 @@
     }
   },
   "/content_types/44dcc7493c65": {
-    "name": "SafetyConcept",
+    "name": "SafetyConsiderations",
     "active": true,
     "id": "44dcc7493c65",
     "contentNodes": {
